requirements.txt
jinja2
numpy
adafruit-circuitpython-servokit
adafruit-circuitpython-bno055
tensorflow==1.15.0
rex_gym
pybluez
PyYAML
click